Motor control toegevoegd

+++ include/tofsensor.hpp
@@ -7,7 +7,7 @@
 
 struct TOFSensorData {
     uint16_t    degree;                        // Value for which corner the servo ha turned
-    float       distance;                      // The distance measured by the tof sensor
+    uint16_t    distance;                      // The distance measured by the tof sensor
 };
